documentation created

+++ src/main/java/com/br/compass/model/Product.java
@@ -1,7 +1,5 @@
 package com.br.compass.model;
 
-import java.math.BigDecimal;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -15,12 +13,12 @@ public class Product {
 	private Long id;
 	private String name;
 	private String description;
-	private BigDecimal price;
+	private Double price;
 	
 	public Product() {
 	}
 
-	public Product(String name, String description, BigDecimal price) {
+	public Product(String name, String description, Double price) {
 		this.name = name;
 		this.description = description;
 		this.price = price;
@@ -50,11 +48,11 @@ public class Product {
 		this.description = description;
 	}
 
-	public BigDecimal getPrice() {
+	public Double getPrice() {
 		return price;
 	}
 
-	public void setPrice(BigDecimal price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
 }
